Fix get profile bug (long to int)

+++ server-dietideals24/src/main/java/com/ucm/serverdietideals24/APIController.java
@@ -62,7 +62,7 @@ public class APIController {
 
     @PostMapping("/generate-login-token")
     public ResponseEntity<String> generateLoginToken(@RequestBody UserFromLoginForm loginReq) {
-        int userId = -1;
+        Long userId = -1L;
 
         try {
             userId = jdbcTemplate.query(
